feat: 홈 화면 부문 검색 필터링 기능 추가 및 부문 전체 조회 API 추가, UI 추가
- AppBar를 home 이라는 디렉토리로 옮김
- AppBar -> HomeTopBar 로 컴포넌트 이름 수정
