+++ front-end/legeno-around-here/src/components/pages/home/HomePage.js
@@ -1,27 +1,29 @@
 import React, { useEffect, useState } from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
 
-import AppBar from '../AppBar';
-import Bottom from '../Bottom';
+import HomeTopBar from './HomeTopBar';
+import Bottom from '../../Bottom';
 
-import { findCurrentPostsFromPage } from '../api/API';
-import { getAccessTokenFromCookie } from '../../util/TokenUtils';
-import { HOME } from '../../constants/BottomItems';
-import PostItem from '../post/PostItem';
-import Loading from '../Loading';
-import BottomBlank from '../BottomBlank';
+import { findCurrentPostsFromPage } from '../../api/API';
+import { getAccessTokenFromCookie } from '../../../util/TokenUtils';
+import { HOME } from '../../../constants/BottomItems';
+import PostItem from '../../post/PostItem';
+import Loading from '../../Loading';
+import BottomBlank from '../../BottomBlank';
+import Container from '@material-ui/core/Container';
 
 const HomePage = () => {
   const [page, setPage] = useState(0);
   const [posts, setPosts] = useState([]);
   const [hasMore, setHasMore] = useState(true);
+  const [sectorId, setSectorId] = useState(0);
 
   const accessToken = getAccessTokenFromCookie();
   const mainAreaId = localStorage.getItem('mainAreaId');
 
   /* 처음에 보여줄 최근글 목록을 가져옴 */
   useEffect(() => {
-    findCurrentPostsFromPage(mainAreaId, 0, accessToken).then((firstPosts) => {
+    findCurrentPostsFromPage(mainAreaId, 0, sectorId, accessToken).then((firstPosts) => {
       if (firstPosts.length === 0) {
         setHasMore(false);
         return;
@@ -29,7 +31,7 @@ const HomePage = () => {
       setPosts(firstPosts);
     });
     setPage(1);
-  }, [mainAreaId, accessToken]);
+  }, [mainAreaId, accessToken, sectorId]);
 
   const fetchNextPosts = () => {
     findCurrentPostsFromPage(mainAreaId, page, accessToken)
@@ -49,19 +51,21 @@ const HomePage = () => {
 
   return (
     <>
-      <AppBar />
-      <InfiniteScroll
-        next={fetchNextPosts}
-        hasMore={hasMore}
-        loader={<Loading />}
-        dataLength={posts.length}
-        endMessage={<h3>모두 읽으셨습니다!</h3>}
-      >
-        {posts.map((post) => (
-          <PostItem key={post.id} post={post} />
-        ))}
-      </InfiniteScroll>
-      <BottomBlank />
+      <HomeTopBar setSectorId={setSectorId} />
+      <Container>
+        <InfiniteScroll
+          next={fetchNextPosts}
+          hasMore={hasMore}
+          loader={<Loading />}
+          dataLength={posts.length}
+          endMessage={<h3>모두 읽으셨습니다!</h3>}
+        >
+          {posts.map((post) => (
+            <PostItem key={post.id} post={post} />
+          ))}
+        </InfiniteScroll>
+        <BottomBlank />
+      </Container>
       <Bottom selected={HOME} />
     </>
   );
